Fix warning: implicit conversion changes signedness: 'unsigned int' to 'int' [-Wsign-conversion]
visp/modules/io/src/image/private/vpImageIoPortable.cpp:516:30: warning: implicit conversion changes signedness: 'unsigned int' to 'int' [-Wsign-conversion]
  516 |   for (int i = I.getHeight() - 1; i >= 0; i--) {
      |            ~   ~~~~~~~~~~~~~~^~~
visp/modules/io/src/image/private/vpImageIoPortable.cpp:591:30: warning: implicit conversion changes signedness: 'unsigned int' to 'int' [-Wsign-conversion]
  591 |   for (int i = I.getHeight() - 1; i >= 0; i--) {
      |            ~   ~~~~~~~~~~~~~~^~~

+++ modules/io/src/image/private/vpImageIoPortable.cpp
@@ -1,6 +1,6 @@
 /*
  * ViSP, open source Visual Servoing Platform software.
- * Copyright (C) 2005 - 2025 by Inria. All rights reserved.
+ * Copyright (C) 2005 - 2026 by Inria. All rights reserved.
  *
  * This software is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -513,7 +513,7 @@ void vp_readPFM_HDR(vpImage<float> &I, const std::string &filename)
 #else
   bool swapEndianness = littleEndian;
 #endif
-  for (int i = I.getHeight() - 1; i >= 0; i--) {
+  for (int i = static_cast<int>(I.getHeight()) - 1; i >= 0; i--) {
     fd.read((char *)I[i], sizeof(float) * w * channels);
     if (swapEndianness) {
       for (unsigned int j = 0; j < w * channels; ++j) {
@@ -588,7 +588,7 @@ void vp_readPFM_HDR(vpImage<vpRGBf> &I, const std::string &filename)
 #else
   bool swapEndianness = littleEndian;
 #endif
-  for (int i = I.getHeight() - 1; i >= 0; i--) {
+  for (int i = static_cast<int>(I.getHeight()) - 1; i >= 0; i--) {
     fd.read((char *)I[i], sizeof(float) * w * channels);
     if (swapEndianness) {
       for (unsigned int j = 0; j < w; ++j) {
